# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Red Intelfon - Workflow de Activaciones API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciales inválidas
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Obtener usuario autenticado
      responses:
        "200":
          description: Usuario actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: No autorizado
...
  /etapas/{etapaId}/checklist:
    get:
      operationId: getChecklist
      tags: [checklist]
      summary: Obtener checklist de una etapa
      parameters:
        - name: etapaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /checklist/{itemId}/toggle:
    put:
      operationId: toggleChecklistItem
      tags: [checklist]
      summary: Marcar/desmarcar item de checklist
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistToggleInput"
      responses:
        "200":
          description: Item actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistItem"

  # Chat
  /procesos/{id}/chat/{etapaOrigen}/{etapaDestino}:
    get:
      operationId: getChatMensajes
      tags: [chat]
      summary: Obtener mensajes del chat entre etapas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: etapaOrigen
          in: path
          required: true
          schema:
            type: integer
        - name: etapaDestino
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de mensajes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMensaje"
    post:
      operationId: sendChatMensaje
      tags: [chat]
      summary: Enviar mensaje en chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: etapaOrigen
          in: path
          required: true
          schema:
            type: integer
        - name: etapaDestino
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMensajeInput"
      responses:
        "201":
          description: Mensaje enviado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMensaje"

  # Notificaciones
...
  /admin/usuarios:
    get:
      operationId: listUsuarios
      tags: [admin]
      summary: Listar usuarios
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      operationId: createUsuario
      tags: [admin]
      summary: Crear usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /admin/usuarios/{id}:
    put:
      operationId: updateUsuario
      tags: [admin]
      summary: Editar usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioUpdate"
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
    delete:
      operationId: deleteUsuario
      tags: [admin]
      summary: Eliminar usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Usuario eliminado

  # Admin - Etapas Config
...
  /admin/etapas/{id}:
    put:
      operationId: updateConfigEtapa
      tags: [admin]
      summary: Actualizar configuración de etapa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigEtapaUpdate"
      responses:
        "200":
          description: Configuración actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigEtapa"

  # Admin - SLA
  /admin/sla-config:
    get:
      operationId: getSlaConfig
      tags: [admin]
      summary: Obtener configuración de SLA
      responses:
        "200":
          description: Configuración de SLA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaConfig"
    put:
      operationId: updateSlaConfig
      tags: [admin]
      summary: Actualizar configuración de SLA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlaConfigUpdate"
      responses:
        "200":
          description: Configuración actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaConfig"

  # Admin - Reportes
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, usuario]
      properties:
        token:
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"

    # Usuario
    Usuario:
      type: object
      required: [id, nombre, email, rol, activo]
      properties:
        id:
          type: integer
        nombre:
          type: string
        email:
          type: string
        rol:
          type: string
          enum: [Ventas, Activaciones, Bodega, MSO, Logistica, Admin]
        area:
          type: ["string", "null"]
        activo:
          type: boolean
        fechaCreacion:
          type: string
          format: date-time
    UsuarioInput:
      type: object
      required: [nombre, email, password, rol]
      properties:
        nombre:
          type: string
        email:
          type: string
        password:
          type: string
        rol:
          type: string
          enum: [Ventas, Activaciones, Bodega, MSO, Logistica, Admin]
        area:
          type: string
    UsuarioUpdate:
      type: object
      properties:
        nombre:
          type: string
        email:
          type: string
        rol:
          type: string
        area:
          type: string
        activo:
          type: boolean

    # Proceso
...
    ChecklistItem:
      type: object
      required: [id, idEtapaProceso, descripcion, completado]
      properties:
        id:
          type: integer
        idEtapaProceso:
          type: integer
        descripcion:
          type: string
        completado:
          type: boolean
        notas:
          type: ["string", "null"]
        fechaCompletado:
          type: ["string", "null"]
          format: date-time
        usuarioQuienCompletoId:
          type: ["integer", "null"]
    ChecklistToggleInput:
      type: object
      required: [completado]
      properties:
        completado:
          type: boolean
        notas:
          type: string

    # Chat
    ChatMensaje:
      type: object
      required: [id, idProceso, etapaOrigen, etapaDestino, usuarioRemitenteId, contenido, fechaMensaje, leido]
      properties:
        id:
          type: integer
        idProceso:
          type: integer
        etapaOrigen:
          type: integer
        etapaDestino:
          type: integer
        usuarioRemitenteId:
          type: integer
        contenido:
          type: string
        fechaMensaje:
          type: string
          format: date-time
        leido:
          type: boolean
        nombreRemitente:
          type: string
        rolRemitente:
          type: string
    ChatMensajeInput:
      type: object
      required: [contenido]
      properties:
        contenido:
          type: string

    # Notificaciones
...
    ConfigEtapa:
      type: object
      required: [id, numeroEtapa, nombreEtapa, slaHoras, activa]
      properties:
        id:
          type: integer
        numeroEtapa:
          type: integer
        nombreEtapa:
          type: string
        descripcion:
          type: ["string", "null"]
        slaHoras:
          type: integer
        areasInvolucradas:
          type: array
          items:
            type: string
        checklistTemplate:
          type: array
          items:
            type: object
            properties:
              descripcion:
                type: string
        activa:
          type: boolean
        ordenVisualizacion:
          type: integer
    ConfigEtapaUpdate:
      type: object
      properties:
        nombreEtapa:
          type: string
        descripcion:
          type: string
        slaHoras:
          type: integer
        areasInvolucradas:
          type: array
          items:
            type: string
        checklistTemplate:
          type: array
          items:
            type: object
            properties:
              descripcion:
                type: string
        activa:
          type: boolean

    # SLA Config
    SlaConfig:
      type: object
      required: [id, nombre, slaGlobalHoras, alertaPorcentaje, activo]
      properties:
        id:
          type: integer
        nombre:
          type: string
        slaGlobalHoras:
          type: integer
        alertaPorcentaje:
          type: integer
        activo:
          type: boolean
    SlaConfigUpdate:
      type: object
      properties:
        nombre:
          type: string
        slaGlobalHoras:
          type: integer
        alertaPorcentaje:
          type: integer

    # Reportes
...