# lib/api-zod/src/generated/types/chatMensaje.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Red Intelfon - Workflow de Activaciones API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMensaje {
  id: number;
  idProceso: number;
  etapaOrigen: number;
  etapaDestino: number;
  usuarioRemitenteId: number;
  contenido: string;
  fechaMensaje: Date;
  leido: boolean;
  nombreRemitente?: string;
  rolRemitente?: string;
}