# lib/api-zod/src/generated/types/usuarioInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Red Intelfon - Workflow de Activaciones API
 * OpenAPI spec version: 0.1.0
 */
import type { UsuarioInputRol } from './usuarioInputRol';

export interface UsuarioInput {
  nombre: string;
  email: string;
  password: string;
  rol: UsuarioInputRol;
  area?: string;
}